+++ tests/testthat/test-utilities-data-combined.R
@@ -1,6 +1,8 @@
 skip_on_os("linux")
 skip_on_ci()
 
+tolerance <- 0.0001
+
 # load the simulation
 simFilePath <- system.file("extdata", "Aciclovir.pkml", package = "ospsuite")
 sim <- loadSimulation(simFilePath)
@@ -22,64 +24,88 @@ myDC <- DataCombined$new()
 myDC$addSimulationResults(simData, groups = "myGroup")
 myDC$addDataSets(obsData, groups = "myGroup")
 
-test_that(
-  "NULL passed to the calculateResiduals function is an error",
-  expect_error(
-    calculateResiduals(NULL)
-  )
-)
+test_that("NULL passed to the calculateResiduals function is an error", {
+  expect_error(calculateResiduals(NULL))
+})
 
 test_that(
   "An empty dataCombined object passed to the calculateResiduals function results in an error",
-  expect_error(
-    calculateResiduals(DataCombined$new())
-  )
+  {
+    expect_error(calculateResiduals(DataCombined$new()))
+  }
 )
 
-test_that(
-  "calculateResiduals returns a dataframe",
+test_that("calculateResiduals returns a dataframe", {
   expect_true(
     "data.frame" %in% class(calculateResiduals(myDC, scaling = "lin"))
   )
-)
-expected_column_names <- c("group", "name", "xValues", "xUnit", "xDimension", "yValuesObserved", "yUnit", "yDimension", "yErrorValues", "yErrorType", "yErrorUnit", "lloq", "yValuesSimulated", "residualValues")
-test_that(
-  "calculateResiduals returns expected columns",
+})
+
+
+
+test_that("calculateResiduals returns expected columns", {
+  expected_column_names <- c(
+    "group",
+    "name",
+    "xValues",
+    "xUnit",
+    "xDimension",
+    "yValuesObserved",
+    "yUnit",
+    "yDimension",
+    "yErrorValues",
+    "yErrorType",
+    "yErrorUnit",
+    "lloq",
+    "yValuesSimulated",
+    "residualValues"
+  )
+
   expect_setequal(
     expected_column_names, colnames(calculateResiduals(myDC, scaling = "lin"))
   )
-)
+})
 
 test_that(
   "DataCombined objects keep LLOQ data passed from underlying DataSet objects",
-  expect_equal(
-    myDC$toDataFrame()$lloq,
-    c(rep(NA, nrow(df)), rep(0.02, length(obsData$yValues)))
-  )
+  {
+    expect_identical(
+      myDC$toDataFrame()$lloq,
+      c(rep(as.numeric(NA), nrow(df)), rep(0.02, length(obsData$yValues))),
+      tolerance = tolerance
+    )
+  }
 )
 
 test_that(
   "calculateResiduals function keeps passes lloq data through",
-  expect_equal(
-    calculateResiduals(myDC, scaling = "lin")$lloq,
-    rep(0.02, 6)
-  )
+  {
+    expect_equal(
+      calculateResiduals(myDC, scaling = "lin")$lloq,
+      rep(0.02, 6),
+      tolerance = tolerance
+    )
+  }
 )
 
 test_that(
   "calculateResiduals does not return rows for data outside of the simulation time",
-  expect_equal(
-    nrow(calculateResiduals(myDC, scaling = "lin")),
-    6
-  )
+  {
+    expect_equal(
+      nrow(calculateResiduals(myDC, scaling = "lin")),
+      6
+    )
+  }
 )
 
 test_that(
   "calculateResiduals returns a correct vector of linear residuals on example data",
-  expect_equal(calculateResiduals(myDC, scaling = "lin")$residualValues,
-    c(0, 0.1000000, -0.0999999, 0.0, -0.1999998, -1),
-    tolerance = 1e-5
-  )
+  {
+    expect_equal(calculateResiduals(myDC, scaling = "lin")$residualValues,
+      c(0, 0.1000000, -0.0999999, 0.0, -0.1999998, -1),
+      tolerance = 1e-5
+    )
+  }
 )
 
 test_that(