+++ tests/testthat/test-utilities-path.R
@@ -1,4 +1,4 @@
-context("toPathArray")
+# toPathArray
 
 test_that("It should convert a valid path to array", {
   expect_identical(toPathArray("Organism|Liver"), c("Organism", "Liver"))
@@ -9,7 +9,7 @@ test_that("It should throw an error if the argument is not a string ", {
 })
 
 
-context("toPathString")
+# toPathString
 
 test_that("It should convert a valid path array to string", {
   expect_identical(toPathString("Organism", "Liver"), "Organism|Liver")